pyproject.toml
[tool.poetry]
name = "hashtagapp-api"
version = "0.1.0"
description = "REST api for basic income hashtag app"
authors = ["prasanta <prasantakakati1994@gmail.com>"]
license = ""

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
django-rest-framework = "^0.1.0"
uwsgi = "^2.0"
psycopg2-binary = "^2.8"
requests = "^2.22"
requests_oauthlib = "^1.2"
tweepy = "^3.8"
pillow = "^6.2"
django-cors-headers = "^3.1"

[tool.poetry.dev-dependencies]
pep8 = "^1.7"
autopep8 = "^1.4"
ipython = "^7.8"
flake8 = "^3.7"
python-language-server = {version = "^0.28.3", extras = ["all"]}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
